requirements.txt
setuptools~=59.6.0
numpy~=1.22.3
PyQt5~=5.15.6
pyqtgraph~=0.12.3
pillow~=9.0.1
uedinst>=1.3.3
tqdm~=4.63
hdf5plugin~=3.3.1
h5py~=3.7.0
pytango~=9.4.1
